working GDI with custom buffer

+++ universalui/universalui_types/src/graphics/mod.rs
@@ -1,10 +1,10 @@
 
 
 pub struct uPixel {
-    pub red: f32,
-    pub green: f32,
-    pub blue: f32,
-    pub alpha: f32
+    pub red: u8,
+    pub green: u8,
+    pub blue: u8,
+    pub alpha: u8
 }
 
 pub struct uPixelBuffer {
